+++ Tests/AgendaTests.swift
@@ -12,80 +12,38 @@ import XCTest
 class AgendaTests: XCTestCase {
 
     private func loadItems(fromFileNamed fileName: String) throws -> (items: [AgendaItem], pagination: AgendaParser.Pagination) {
-        let path = try XCTUnwrap(Bundle(for: type(of: self)).path(forResource: fileName, ofType: "html"))
-        let html = try XCTUnwrap(String(contentsOfFile: path))
-        let items = try XCTUnwrap(AgendaParser.parseItems(html: html))
+        let path = try XCTUnwrap(Bundle(for: type(of: self)).path(forResource: fileName, ofType: "rss"))
+        let rss = try XCTUnwrap(String(contentsOfFile: path))
+        let items = try XCTUnwrap(AgendaParser.parseItems(rss: rss))
         return items
     }
 
-    func testParseAgendaItems1() throws {
-        let parsedObjects = try loadItems(fromFileNamed: "TestAgenda1")
+    func testParseRSSFeed() throws {
+        let parsedObjects = try loadItems(fromFileNamed: "TestAgenda-2024-1")
 
-        XCTAssertNotNil(parsedObjects.pagination.nextPage)
-        let nextPageURL = try XCTUnwrap(parsedObjects.pagination.nextPage)
-        XCTAssertEqual(nextPageURL.absoluteString, "https://www.bm-grenoble.fr/688-agenda.htm?TPL_CODE=TPL_AGENDALISTE&ip=2&op=AGP_DATEFIN+asc&cp=998b0016113da5c36d2a&mp=10#p")
+        XCTAssertNil(parsedObjects.pagination.nextPage)
+        // let nextPageURL = try XCTUnwrap(parsedObjects.pagination.nextPage)
+        // XCTAssertEqual(nextPageURL.absoluteString, "https://www.bm-grenoble.fr/688-agenda.htm?TPL_CODE=TPL_AGENDALISTE&ip=2&op=AGP_DATEFIN+asc&cp=998b0016113da5c36d2a&mp=10#p")
 
         let items = parsedObjects.items
-        XCTAssertEqual(items.count, 10)
+        XCTAssertEqual(items.count, 2)
 
-        XCTAssertEqual(items[0].title, "Le nudge, comment influencer sans contrainte")
-        XCTAssertEqual(items[0].category, "Conférences")
+        XCTAssertEqual(items[0].title, "Parcoursup : les algorithmes et leurs impacts")
+        XCTAssertEqual(items[0].category, "Conférence - débat")
         XCTAssertEqual(items[0].library, "Bibliothèque Kateb Yacine")
-        XCTAssertEqual(items[0].summary, "Une conférence animée par Evelyn Rosset, docteur en psychologie, chercheure associée au LIP/PC2S, UGA.")
-        XCTAssertEqual(items[0].link.absoluteString, "https://www.bm-grenoble.fr/688-agenda.htm?TPL_CODE=TPL_AGENDA&PAR_TPL_IDENTIFIANT=7863")
+        // XCTAssertEqual(items[0].summary, "Une conférence animée par Evelyn Rosset, docteur en psychologie, chercheure associée au LIP/PC2S, UGA.")
+        XCTAssertEqual(items[0].link.absoluteString, "https://bm-grenoble.fr/Default/doc/CALENDAR/40/parcoursup-les-algorithmes-et-leurs-impacts")
         let items0Image = try XCTUnwrap(items[0].image)
-        XCTAssertEqual(items0Image.absoluteString, "https://www.bm-grenoble.fr/uploads/Image/d0/IMF_TETIERE/25362_085_1h-de-psy-2019.jpg")
+        XCTAssertEqual(items0Image.absoluteString, "https://bm-grenoble.fr/basicimagedownload.ashx?itemGuid=DB23C569-C01C-45B7-9E83-F78A2A04CCD6")
         switch items[0].date {
         case .range:
             XCTFail()
         case .day(let dateComponents):
-            XCTAssertEqual(dateComponents.day, 11)
-            XCTAssertEqual(dateComponents.month, 1)
-            XCTAssertEqual(dateComponents.year, 2020)
-        }
-
-        XCTAssertEqual(items[3].library, "Bibliothèque Eaux-Claires Mistral")
-        XCTAssertEqual(items[6].summary, "Un programme de lectures ponctuées de chansons sur une musique originale de Thierry Ronget.")
-    }
-
-    func testParseAgendaItems2() throws {
-        let parsedObjects = try loadItems(fromFileNamed: "TestAgenda2")
-
-        XCTAssertNotNil(parsedObjects.pagination.nextPage)
-        let nextPageURL = try XCTUnwrap(parsedObjects.pagination.nextPage)
-        XCTAssertEqual(nextPageURL.absoluteString, "https://www.bm-grenoble.fr/688-agenda.htm?TPL_CODE=TPL_AGENDALISTE&ip=3&op=AGP_DATEFIN+asc&cp=998b0016113da5c36d2a&mp=10#p")
-
-        let items = parsedObjects.items
-        XCTAssertEqual(items.count, 10)
-
-        switch items[2].date {
-        case .day:
+            XCTAssertEqual(dateComponents.day, 9)
+            XCTAssertEqual(dateComponents.month, 2)
+            XCTAssertEqual(dateComponents.year, 2024)
+        case .none:
             XCTFail()
-        case .range(let startDateComponents, let endDateComponents):
-            XCTAssertEqual(startDateComponents.day, 6)
-            XCTAssertEqual(startDateComponents.month, 12)
-            XCTAssertEqual(startDateComponents.year, 2019)
-
-            XCTAssertEqual(endDateComponents.day, 14)
-            XCTAssertEqual(endDateComponents.month, 2)
-            XCTAssertEqual(endDateComponents.year, 2020)
         }
-    }
-
-    func testParseAgendaItems3() throws {
-        let parsedObjects = try loadItems(fromFileNamed: "TestAgenda3")
-        XCTAssertNil(parsedObjects.pagination.nextPage)
-    }
-    
-    func testParseAgendaItems4() throws {
-        let parsedObjects = try loadItems(fromFileNamed: "TestAgenda4")
-        let items = parsedObjects.items
-        XCTAssertEqual(items.count, 10)
-        
-        let library0 = try XCTUnwrap(items[0].library)
-        XCTAssertEqual(library0, "")
-        
-        let library = try XCTUnwrap(items[1].library)
-        XCTAssertEqual(library, "")
     }
 }