+++ bm/Agenda/AgendaItem.swift
@@ -15,11 +15,11 @@ struct AgendaItem: Codable {
     }
 
     let title: String
-    let summary: String
-    let category: String
+    let summary: String?
+    let category: String?
     let library: String?
     let link: URL
-    let date: AgendaDate
+    let date: AgendaDate?
     let image: URL?
 }